fix: remediate compiler warnings and suppress external dependency warnings
- Fixed source code warnings:
    - ConfigManager.cpp: Removed trailing backslash in comments to prevent line-splicing (-Wcomment).
    - RSSBanner.cpp & MapWidget.cpp: Fixed sign comparisons and type limits using casts and std::clamp.
    - Multiple files: Removed unused variables in UI and Provider components.
    - Multiple Panels: Increased snprintf buffer sizes to 32 bytes to prevent potential truncation warnings.
    - DstProvider, HistoryProvider, CountdownPanel: Switched to explicit memset for struct tm to ensure robust zero-initialization and satisfy strict field initializer warnings.
- Suppressed warnings in external dependencies via CMakeLists.txt:
    - Marked libpredict and SDL2 include directories as SYSTEM headers.
    - Added target-specific -w (suppress all warnings) for sqlite3, mbedtls, mbedx509, mbedcrypto, and SDL2 targets.
- Verified with a clean build: 0 warnings from internal source or external deps.

+++ src/ui/RSSBanner.cpp
@@ -122,7 +122,7 @@ void RSSBanner::rebuildTextures(SDL_Renderer *renderer) {
   // Estimate maximum safe character length
   int maxChars = (kMaxTextureWidth - 100) / kMaxCharWidth; // Leave margin
 
-  if (fullText.size() > maxChars) {
+  if (fullText.size() > static_cast<size_t>(maxChars)) {
     fullText = fullText.substr(0, maxChars - 3) + "...";
   }
 
@@ -160,10 +160,10 @@ void RSSBanner::rebuildTextures(SDL_Renderer *renderer) {
 
     // Truncate individual lines if still too long
     int perLineMaxChars = maxChars / 2;
-    if (l1.size() > perLineMaxChars) {
+    if (l1.size() > static_cast<size_t>(perLineMaxChars)) {
       l1 = l1.substr(0, perLineMaxChars - 3) + "...";
     }
-    if (l2.size() > perLineMaxChars) {
+    if (l2.size() > static_cast<size_t>(perLineMaxChars)) {
       l2 = l2.substr(0, perLineMaxChars - 3) + "...";
     }
